raw/requirements.txt
chardet==5.2.0
colorama==0.4.6
lazy-object-proxy==1.10.0
piexif==1.1.3
pillow==11.0.0
